PB/requirements.txt
APScheduler==3.9.1.post1
asgiref==3.5.2
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
googlemaps==4.6.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
phonenumbers==8.12.57
Pillow==9.3.0
pycodestyle==2.9.1
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.17.6
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.6
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
